evmscan migration: migrate to use evmscan to support other major EVM-based chains
Now it supports BSC, Ethereum, and Polygon for the input address.

src/types.rs, and src/util.rs are added to help the logic code, and
abstract stuff better.

Brought in clap, and regex crate as well to help in validating cli's
arguments, and utility function to check validity of input address.